Code reorg to build the plugin lib
