+++ Cargo.toml
@@ -26,11 +26,12 @@ description = "Zenoh plugin for ROS2 and DDS in general"
 
 [lib]
 name = "zplugin_dds"
-# crate-type = ["cdylib"]
+crate-type = ["cdylib", "rlib"]
 
 
 [dependencies]
 zenoh =  { git = "https://github.com/eclipse-zenoh/zenoh"}
+zenoh-plugin-rest =  { git = "https://github.com/eclipse-zenoh/zenoh"}
 cyclors = "0.1.1"
 log = "0.4.14"
 async-trait = "0.1.42"
